feat: map and sidebar spli + transitions
