+++ src/app/globals.css
@@ -1,6 +1,10 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+*{
+  /* outline:1px solid red; */
+}
  
 @layer base {
   :root {
