+++ src/components/Maplibre/useMaplibreMap.ts
@@ -5,8 +5,17 @@ function useMapboxMap(options: any) {
   const [map, setMap] = useState<any | null>(null);
   const [maploaded, setMaploaded] = useState(false);
   const mapRef = useRef<HTMLDivElement | null>(null);
+  
+  console.log(
+    "🚀 ~ file: useMaplibreMap.ts:8 ~ useMapboxMap ~ mapRef:",
+    mapRef
+  );
 
   useEffect(() => {
+    if (!mapRef.current) {
+      console.error("Refrense to map not provided");
+      return;
+    }
     const opts = {
       container: mapRef.current,
       preserveDrawingBuffer: true,
@@ -21,7 +30,7 @@ function useMapboxMap(options: any) {
               // 'https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}',
             ],
             tileSize: 256,
-        
+
             maxzoom: 18,
           },
         },
